refactor(cyclonedx): add marshaler
